Input, scalling, printing and stuff
The scalling is now working when adjusting the terminal size
Also got a bit of scrolling to work (Down, Left, Right)
Changing category is also working(Left, Right)
Real-time input taking is also added, though I haven't use it to match or done anything(most likely will in the next commit)

The code or the logic to get these working might seem a lil weird, but it works.
Also the program refreshes every 10 miliseconds, I am not sure if this will affect the performance or cause problems, so I might need to change the rate itf.