Okay
The scrolling up and down with the pointer work so that's nice
Here's the fucking problem: I DON'T KNOW HOW TO GET THE POINTER TO THE END OF THE VECTOR WITH THE SCALLING STILL WORKING
maybe it's my fault that the scalling system is shit and it really is.
Ok I need to calm down first and then I'll try my best to fix the scalling and the pointer movements
Wish me luck! (pls)