The UI
I've finally added the main function of the program copy this supports both x11 and wayland
And the UI just looks so fricking good right now, I'm about to cry. Yes I've done the UI but it will need a bit of changing later(User Manual)
The method used to make the UI is pretty striaght forward by comparing the terminal size and print according to it
The UI is also very scallable, you could adjust the terminal how much you want and it would still look nice!
This project is finally coming to life >;3